+++ app/src/main/java/woowacourse/shopping/ui/component/productlist/ProductQuantityAction.kt
@@ -1,0 +1,39 @@
+package woowacourse.shopping.ui.component.productlist
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import woowacourse.shopping.ui.theme.AndroidShoppingTheme
+
+@Composable
+fun ProductQuantityAction(
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = modifier.clickable(onClick = onClick),
+    ) {
+        Text(
+            text = text,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+        )
+    }
+}
+
+@Composable
+@Preview(showBackground = true)
+private fun ProductQuantityActionPreview() {
+    AndroidShoppingTheme {
+        ProductQuantityAction(
+            text = "+",
+            onClick = {},
+        )
+    }
+}
