+++ app/src/main/java/woowacourse/shopping/ui/component/productlist/ShoppingCardAddBox.kt
@@ -1,6 +1,6 @@
 @file:Suppress("FunctionName")
 
-package woowacourse.shopping.ui.component
+package woowacourse.shopping.ui.component.productlist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
